Move analyticsDatabase.ts
